refactor :: rename 404 to NotFound
